Extract ProjectLocatorSpec
